Synced example study test utils and test-addon test utils + split test utils into separate files based on type of test util method
